ap51-flash:  separate main() from ap51-flash.c into main.c
The Makefile now creates an archive containing all .o files and link against
this archive, instead of each .o file on its own. This will make developers
working on external projects, that want to provide the ap51-flash features,
more happy. They can copy the archive and necessary header-files and just
include and link against them.

A static library can be created more easily, because of the separation
of the main() function into a separate file.

ap51-flash.h and libap51-flash.a need to be exported into the external
project and can then be used.

Signed-off-by: Julian <julian@svg4all.de>
Signed-off-by: Marek Lindner <lindner_marek@yahoo.de>