ap51-flash: Use fwupgrade.cfg endings as flexible image TYPE
CE images have a limited size field for the TYPE. This is usually something
like OM5P or MR900. But combined images can end up with things like
"OM2P,OM5P,OM5PAC,MR600,MR900,MR1750". This is too long and thus cannot be
used. But it is possible to use the ending of the fwupgrade.cfg files to detect
which devices are supported and avoid this size probem.

The size problem is not completely fixed because only up to 255 files are
supported in a CE image format. And each filename can only be 32 bytes long.

Signed-off-by: Sven Eckelmann <sven@open-mesh.com>