ap51-flash: Disallow load new images for already loaded router_images
An router_image can only store the data for a single external or embedded
image. All offsets and sizes are only for this single image. Trying to load
a different image with the same type will result in a new path for this
router_image but the offsets and sizes for the old image is still stored in
ap51-flash. This leads to errors during the flash process.

Signed-off-by: Sven Eckelmann <sven@open-mesh.com>
Signed-off-by: Marek Lindner <mareklindner@neomailbox.ch>