ap51-flash: initialize embedded images after commandline images
The ap51-flash tool with embedded images cannot be used with external
images of the same type as the embedded images because these already use
the router_image object for this type.

Instead load the embedded images after the external images and only when no
other image with this type was already loaded.

Signed-off-by: Sven Eckelmann <sven@open-mesh.com>
Signed-off-by: Marek Lindner <mareklindner@neomailbox.ch>