ap51-flash: Don't add fwupgrade.cfg* to the image file_size
An CE multi-device image can have multiple fwupgrade.cfg files which have
different sizes. These different sizes cannot be represented in the
router_image->file_size because it is shared for all routers in the
ce_image.

Also fwupgrade.cfg.sig is a file which is not counted correctly because
some routers may need this file and others can be flashed without this
file.

Instead ignore fwupgrade.cfg for the file_size because it is not the actual
payload but control information. Not counting in while parsing the image
and while transferring the data resolves the problem that the "finish"
messages are not shown correctly for single-device CE images.

Signed-off-by: Sven Eckelmann <sven@open-mesh.com>