ap51-flash: Calculate size of ce images using fwupgrade.cfg
Each device in a multi-device CE image has an fwupgrade.cfg. This
fwupgrade.cfg can be parsed to determine which file actually belongs to
which device. These files are expected to be requested by the device u-boot
and their accumulated size is the expected size to be transferred. The
flash process can end for a device when this number of bytes were
transferred successfully.

Parsing fwupgrade.cfg is required because not all files stored in a
multi-device CE image are for all routers supported by this image. A router
will therefore not always request all files stored in the image.

Signed-off-by: Sven Eckelmann <sven@open-mesh.com>