ap51-flash: make: Mark non-output targets as PHONY
make expects that the name of a target is generated by the rules for this
target. It will therefore not start the all/clean target when there is a
file with the same name which is newer than all the dependencies of this
target.

The .PHONY prevents this check and always assumes that this target has to
be started without looking for already existing files with the same name.

Signed-off-by: Sven Eckelmann <sven@narfation.org>