ap51-flash: Reduce number of misaligned accesses during packet parsing
The ethernet header is only 14 bytes long and thus creates a misalignment
in all the 4-byte aligned headers which come after it. This includes things
like the IP addresses. Access to all these structures causes faults which
have to be handled by the kernel.

This can mostly be resolved by moving the packet header by 2 bytes from the
4-bytes boundary. This will then automatically fix the IP header alignment.

Signed-off-by: Sven Eckelmann <sven.eckelmann@openmesh.com>