ap51-flash: Fix ARP misaligned read/write
The ARP header for IPv4 always have one of the IP addresses not aligned to
a 4-byte boundary. It is necessary to access this memory only with 1 or 2
byte read/writes. Otherwise, architectures without unaligned memory access
support must use the kernel to handle errors during the access of these
memory regions.

Signed-off-by: Sven Eckelmann <sven.eckelmann@openmesh.com>