ap51-flash: Fix out-of-write on packet receive
The packet buffer gets \0 terminated after it was read. The \0 always comes
after the read data. But if we read enough data to fill the buffer, the \0
will be written 1 byte after the buffer ended.

This can be avoided by only reading packet_buff_len - 1 bytes. The \0 will
then only be added at the end of the buffer or any other byte before that.

Signed-off-by: Sven Eckelmann <sven.eckelmann@openmesh.com>