ap51-flash: add support for "netconsole" flashing mode
some newer uboots support the so called "Network Console",
which is basically the uboot serial console exposed via
UDP on port 6666.

A router booted in this mode will start sending the output
of the console to port 6666 of the remote host and will wait
for input on its own port 6666.

To flash a router, ap51-flash has to wait for the console
prompt to appear and send the "fw_upg" command to start
the actual TFTP client session.

Once flashing is over, a 'reset' command must be explicitly
sent to reboot the router.

Some details can be found at[1].

[1]https://github.com/pepe2k/u-boot_mod#network-console

Signed-off-by: Antonio Quartulli <a@unstable.cc>