ap51-flash: netconsole - concatenate reset command to reflash
It seems that the reset command sent at the end of the flashing
procedureeis not received by the router netconsole.

To make the procedure more robust, concatenate 'reset'
to the 'fw_upg' command being sent right at the beginning.

This way the uboot shell will simply execute the two
instructions one after the other.

Signed-off-by: Antonio Quartulli <a@unstable.cc>