ap51-flash: Merge code for printing and retrieving interfaces
The code to retrieve the list of all interfaces and the code to find the
correct interface name based on the index is mostly the same code. Only the
action which is performed for each interfaces is different for these two
functions.

Instead of adding the code multiple times, just extract the different,
inner code via function pointers. This will also make it easier later to
process multiple RTNL messages from the kernel.

Signed-off-by: Sven Eckelmann <seckelmann@datto.com>