ap51-flash: Use system interface index under linux
The pcap library doesn't retrieve the system interface index. It is
therefore required to create a logical number as index. But the RTNL based
code is getting this information directly from the kernel.

It is safer to use the global system index in most situation because it is
not affected by reordering and renumbering - something which could happen
with libpcap when unrelated interfaces are added/removed.

Signed-off-by: Sven Eckelmann <seckelmann@datto.com>