ap51-flash: Parse RTNL packet before closing netlink socket
The RTNL response packet for RTM_GETLINK was parsed after the socket is
closed. This would work when all RTNL messages are stored in a single
packet. This is not the case and multiple messages will be send.

Parsing the packet before the socket is closed will allow ap51-flash to
parse multiple packets in the future.

Signed-off-by: Sven Eckelmann <seckelmann@datto.com>