ap51-flash: Fix retry of RTNL message download
The expected message size is also the size of our allocated buffer. This
size is first retrieved via the peek into the buffer. And it is then given
as size argument to the actual recv call. And this actual recv call can be
called multiple times in case the system call was interrupted.

The actual receive call must therefore not overwrite the size information
or otherwise it will fail to download the correct data on the second try.

Fixes: e9f2088bd9dc ("ap51-flash: Retry RTNL reply recv after interruption")
Signed-off-by: Sven Eckelmann <sven@narfation.org>