ap51-flash: Test Npcap SDK builds for delayed DLLs
The modern Npcap DLLs are the only ones which are installed by default by
Npcap. The WinPCAP compatibility DLLs are only installed as optional
feature.

Unfortunately, the modern Npcap DLLs are in a Windows subdirectory which is
not searched by default. The program must instead take care of selecting
the new path and thus load the libraries in a delayed fashion.

Signed-off-by: Sven Eckelmann <sven@narfation.org>