ap51-flash: Use direct double linked list for objects
The old list implementation was less abstract and involved manual fiddling
with pointers. It was also using an extra object to store the list itself -
which required more error handling.

Just keeping the list pointers inside the list reduces the complexity and
makes everything easier to ready with the available helper code.

Signed-off-by: Sven Eckelmann <sven@narfation.org>