ap51-flash: Fix embedding of images via objdump with GCC
The old way of accessing the size information of a variable didn't work
because the size itself is created by objdump as ABS. When relocation is
used, this will cause an error in the linker like:

  /usr/bin/ld: router_images.o: relocation R_X86_64_PC32 against absolute symbol `_binary_img_ci_size' in section `.text' is disallowed
  collect2: error: ld returned 1 exit status

This happens because the GOT relocation is no longer produced since
https://gcc.gnu.org/git/?p=gcc.git&a=commit;h=77ad54d911dd7cb88caf697ac213929f6132fdcf
("x86-64: Optimize access to globals in PIE with copy reloc") and PIE was
enabled on many systems by default (to hardden the binaries by default).

But the size information can simply be calculated by subtracting the start
from the end of the section.

Signed-off-by: Sven Eckelmann <sven@narfation.org>