ap51-flash: convert detect_main packet_buff to Ethernet header
The current device detection code assumes a detection based
on the ARP header. Because of this, the individual detect_main
callback is passed the ARP header as argument.

In order to support detection methods beyond ARP, the
detect_main invocation is modified to pass the Ethernet
header regardless of the upper layer protocol. The check for
ARP has to be handled by the existing detect_main callbacks.

Signed-off-by: Marek Lindner <mareklindner@neomailbox.ch>