ap51-flash: netconsole: Split private and node state
The netconsole code has its own internal state which is basically an
embedded state machine in the global node state machine. While this code
was refactored to simplify it, the code to switch the node state
was accidentally dropped and instead it was tried to save the node state in
the netconsole specific state.

This had no effect until now because the NODE_STATUS_REBOOTED (6) is
outside the possible states for the netconsole. And the code handled this
as "nothing to do".

Fixes: 7a4ef891a229 ("ap51-flash: netconsole - concatenate reset command to reflash")
Signed-off-by: Sven Eckelmann <sven@narfation.org>