ap51-flash: Fix maximum size of fwupgrade.cfg file
A filename inside an CE01 image can be up to 32. This is a hard limit from
the index. The FILE_NAME_MAX_LENGTH is therefore set to 33 (32 characters +
\0 terminator) - exactly like the size of file_name_buff used to calculate
the right name for the fwupgrade.cfg in a "COMBINED" CE01 image.

But the snprintf was set to only be allowed to write 32 characters
(including the \0) to this buffer. So not all possible fwupgrade-... names
could be found by this router_image_get_file.

Fixes: 33b449227163 ("ap51-flash: load platform dependent fwupgrade.cfg first")
Signed-off-by: Sven Eckelmann <sven@narfation.org>