ap51-flash: Ignore overlong generated filenames
For "COMBINED" CE01 images for multiple devices, the standard filenames
"fwupgrade.cfg" and "fwupgrade.cfg.sig" are mapped to device specific
names. This could for example be:

* OM2P: fwupgrade.cfg.sig -> fwupgrade.cfg-OM2P.sig
* A40: fwupgrade.cfg -> fwupgrade.cfg-A60

But for really long router type names, it doesn't work. A filename in the
CE01 index can only be 32 characters long. So the router_type
"REALLYLARGEDEVNAME" would try to calculate the fwupgrade.cfg.sig name
"fwupgrade.cfg-REALLYLARGEDEVNAME.sig". But in the buffer, there is only
enough space for "fwupgrade.cfg-REALLYLARGEDEVNAME". Which is unfortunately
not the name of the fwupgrade.cfg.sig but for the "fwupgrade.cfg.

Just ignore the generated names in case they couldn't be written correctly
in the buffer (completely).

Signed-off-by: Sven Eckelmann <sven@narfation.org>